refactor(window): extract viz modes from music-player to separate module
4 VizMode factories (bars, rings, grid, rain) + VizColors/VizMode
interfaces + VIZ_BANDS constant + VIZ_MODES registry moved to
music-player-viz.ts. music-player-window.ts: 1227 → 848 LOC.
